[breakpoint] Down properties are now inclusive (#9632)
* [Hidden] down properties are now inclusive
xsDown, smDown, mdDown, lgDown, and xlDown match the behavior of their *Up* counterparts and are now inclusive of the specified breakpoint.  This is consistent between the CSS and JS implementation of this component.

Additionally, `theme.breakpoints.down` has been altered to match this behavior.  This function is now consistent with its `up` and `only` counterparts.

Documentation has been updated, as well as internal dependence on the `down` function.  The implementation of isWidthDown has been altered so that its default behavior is to be inclusive (this was hinted in an associated comment, but was not the case).

The decision to move forward with this PR was made following a conversation with @oliviertassinari and @rosskevin, which resulted in a conclusion that it would be better to deal with this inconsistency and the breaking changes its solution would require while we are in beta.

These are the following breaking changes:

- 'createBreakpoints.down' is now inclusive of the specified breakpoint
- `isWidthDown` is now inclusive of the specified breakpoint by default
- `Hidden` will include the breakpoints associated with its *Down* properties regardless of whether CSS or JS is used.

* [Hidden] address PR notes
- Updated hidden doc
- Added an isHidden test
- createBreakpoints down will use the key if it is an unspecified breakpoint
- cleaned up createBreakpoints tests

* [Hidden] API docs update

* [Hidden] fix demos
- AppDrawer and ResponsiveDrawer were using *Down properties that needed to be shifted down one breakpoint due to the changes in this PR
- ResponsiveDrawer was reported in Argos
- AppDrawer was not reported in Argos
- BreakpointDown was reported as a regression, but the current behavior is as expected

* [Hidden] bump .size-limit for overhead to 24.4 from 24.3