[core] Add eslint-plugin-react-hooks (#14629)
* [docs] SimpleSelect cleaner hooks

* [core] Add eslint-plugin-react-hooks

* [docs] Fix broken CircularDeterminate hooks demo

* [docs] Fix leaky CircularIntegration demo

* [docs] Linter appeasement

* [useMediaQuery] Linter appeasment

* [useMediaQuery] Linter appeasment

Arguable if this is actual linter appeasment. The statement
about is not valueable. effects are not exectured on the server
anyway so there is no performance gain. only noise in the code

* [useMediaQuery] Remove redundant bailout

* [useMediaQuery] perf optimization for repeated renders

window.matchMedia was called on every render for defaultValue. We now
only call it on state initialization.

* [makeStyles] Lint appeasment

* [withStyles] Linter appeasement

* [docs] Fix broken LinearDeterminate demo

* [docs] Linter appeasment

Makes for cleaner architecture though. Previously `state`
mixed input values and labelWidth. Now the concerns
are clearly separated

* [docs] Fix LinearBuffer demo

Same approach as
https://overreacted.io/making-setinterval-declarative-with-react-hooks/

* [docs] Add warning about useWidth demo

* [docs} Fix broken scroll logic

Closes #14628

* [TrapFocus] Fix potential bug when props are dynamic

* [makeStyles] Disable hook rules

I suspect that there is a bug related to hooks usage.
Need to investigate